1.1.0 Fixed strong typing and added more documentation (#1)
* docs(package.py): Fixing up how the documentation is generated, making it more robust in case of errors

Also fixed changelog to not appear completely in the documentation-toc
(by not having headings, but rather bold font)

* refactor(value_split): Renamed value_split to path_split, some documentation for that new parameter

* perf(poetry.lock): Updated the package dependencies (poetry update)

* fix(fagus.py,README.md): Added documentation for list_insert, and fixed a bug that was found in the doctests

This bug is now properly tested for by the doctest.

* docs(README): Finally finished documenting the fagus-options

This time to be precise node_types was documented, some other minor changes in the code.

* build(poetry-update): Updating to the latest dependencies, adding mypy as a dev-dependency

* style(mypy---strict): Started to implement mypy --strict in filters and utils

* refactor(mypy---strict): iterators.py

* style(fagus.py): mypy --strict in fagus.py

* feat(ALL): mypy --strict done on the whole project

* style(ALL): mypy --strict compatible with python 3.6

Had to import all the container classes like MutableMapping etc. twice - once through typing which supports MutableMapping[Any, Any] and then again through collections.abc, as only collections.abc properly supports isinstance({}, MutableMapping). Up until Python 3.8, MutableMapping from typing has the type _GenericAlias, which doesn't work with isinstance. So to get the best of both worlds, in all the type hints typing.MutableMapping is used to support MutableMapping[Any, Any], and otherwise c_abc.MutableMapping is used for the code to still work. Cumbersome, but at least a solution. import collections.abc as c_abc can be removed when support for Python < 3.9 is removed

* Docstring for FagusOption

* Update README.md

* fixing the basic doc

* new documentation for list indices

* adding proper docstring for FagusOption

* documented append and set() properly

* test(tests): Fixing doctests, adding mypy to pre-commit-config, updating dependencies

Some smaller changes in files due to automatic whitespace trailing etc.

* feat(update()): Greatly simplified the function using try catch

It now also supports e.g. dict.items() as input, as it will try to update the dict first, and only if that doesn't work a set is created

* docs(update): Fixed documentation for the update function, and the update function itself

* docs(CHANGELOG): Bumped version number to 1.1.0, and explained changes in CHANGELOG

---------

Co-authored-by: Lukas Neuenschwander <lukas@Lukas-MacBook-Air.local>
Co-authored-by: Neuenschwander <lukas.neuenschwander@nbim.no>